backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
python-dateutil==2.8.2

# Kafka
aiokafka==0.10.0

# OpenSearch
opensearch-py==2.4.2

# ML
scikit-learn==1.4.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
